fix vc2010 compile issues
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2925 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ primitive.cpp
@@ -77,7 +77,7 @@ bool Mesh::LoadAnimation(const char *fname, const bool flipTV, const bool conver
       }
 
    }
-   sname = to_string(frameCounter)+" frames imported!";
+   sname = to_string((long long)frameCounter)+" frames imported!";
    MessageBox(NULL, sname.c_str(), "Info", MB_OK | MB_ICONEXCLAMATION);
    return true;
 }
